# openapi/OpenApiSpectralDemo.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "OpenApiSpectralDemo | v1",
    "version": "1.0.0"
  },
  "paths": {
    "/api/WeatherForecast": {
      "get": {
        "tags": [
          "WeatherForecast"
        ],
        "summary": "Retrieves weather forecasts for the next 5 days",
        "operationId": "GetWeatherForecast",
        "responses": {
          "200": {
            "description": "Returns the weather forecasts",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WeatherForecast"
                  }
                }
              }
            }
          },
          "500": {
            "description": "If an internal server error occurs",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/WeatherForecast/{day}": {
      "get": {
        "tags": [
          "WeatherForecast"
        ],
        "summary": "Retrieves a specific weather forecast by day offset",
        "operationId": "GetWeatherForecastByDay",
        "parameters": [
          {
            "name": "day",
            "in": "path",
            "description": "Day offset (1-5)",
            "required": true,
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns the weather forecast",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WeatherForecast"
                }
              }
            }
          },
          "400": {
            "description": "If the day parameter is invalid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "If the forecast is not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "If an internal server error occurs",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "WeatherForecast": {
        "type": "object",
        "properties": {
          "date": {
            "type": "string",
            "description": "The date of the forecast",
            "format": "date"
          },
          "temperatureC": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "description": "Temperature in Celsius",
            "format": "int32"
          },
          "temperatureF": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "description": "Temperature in Fahrenheit",
            "format": "int32"
          },
          "summary": {
            "type": [
              "null",
              "string"
            ],
            "description": "Weather summary description"
          }
        },
        "description": "Weather forecast data model"
      }
    }
  },
  "tags": [
    {
      "name": "WeatherForecast"
    }
  ]
}